fix(text): use long for color property
Flutter uses unsigned int to store color and spring cannot decode it from json, so we must use long

+++ src/main/java/org/fekz115/dip/model/Text.java
@@ -24,6 +24,6 @@ public class Text implements Content {
     private boolean bold;
     private boolean italic;
     private int size;
-    private int color;
+    private long color;
     private boolean separate;
 }
